src/services/Slice/IngredientsSlice.ts
import { util } from 'webpack';
import { getIngredientsApi } from '../../utils/burger-api';
import { TIngredient } from '@utils-types';
import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';

// Типизация
type TIngredientsState = {
  ingredients: TIngredient[];
  loading: boolean;
  error: string | null | undefined;
};
// Асинхронный экшен
export const getIngredients = createAsyncThunk(
  'ingredients/getIngredients',
  async () => getIngredientsApi()
);
//начальное состояние
export const initialState: TIngredientsState = {
  ingredients: [],
  loading: false,
  error: null
};
//Слайс
export const ingredientSlice = createSlice({
  initialState,
  reducers: {},
  name: 'ingredient',
  extraReducers: (builder) => {
    builder
      .addCase(getIngredients.pending, (state) => {
        state.loading = true;
        state.error = null;
      })
      .addCase(getIngredients.fulfilled, (state, action) => {
        state.loading = false;
        state.ingredients = action.payload;
      })
      .addCase(getIngredients.rejected, (state, action) => {
        state.loading = false;
        state.error = action.error.message;
      });
  }
});

/* export default ingredientSlice.reducer; */
